Spark 3 & Hadoop 3 support, Scala 2.13 cross compilation (#3294)
* Upd Spark and Spire, drop Scala 2.11

* Workaround Cassandra JXM issue and commentout a failing test

* Add Scala 2.13 crosscompilation

* Adjust Raster and Spark tests

* Fix Raster tests

* Fix GDAL tests

* Fix HBase tests

* Increase CircleCI no output timeout

* Supply the empty argument list explicitly, it is deprecated since Scala 2.13

* Upd up to Hadoop 3.2.1 and HBase 2.4.2

* Move Accumulo into a separate executor

* Upd SBT up to 1.5.0

* Bump spire up to 0.17.0

* fromRGBA(double) => fromRGBAPct

* Bump proj4j version up

+++ s3-spark/src/main/scala/geotrellis/spark/store/s3/S3RecordReader.scala
@@ -45,7 +45,7 @@ abstract class BaseS3RecordReader[K, V](s3Client: S3Client) extends RecordReader
     logger.debug(s"Initialize split on bucket '$bucket' with $keyCount keys")
   }
 
-  def getProgress: Float = curCount / keyCount
+  def getProgress: Float = (curCount / keyCount).toFloat
 
   def readObjectRequest(objectRequest: GetObjectRequest): (K, V)
